build-requirements.txt
# Build requirements for 影藏·媒体管理器 cross-platform executables
# Install with: pip install -r build-requirements.txt

# Core GUI framework
PySide6>=6.5.0
PySide6>=6.5

# Build tools
pyinstaller>=5.0.0

# Image processing for icon creation (optional)
Pillow>=9.0.0

# TOML parsing for version extraction
tomli>=2.0.0; python_version < "3.11"

# Optional compression tool (smaller executables)
# Note: UPX may not be available on all systems
# Install separately if needed: https://upx.github.io/
# upx>=4.0.0

# Development dependencies (for testing builds)
# Uncomment if needed for testing
# pytest>=7.0.0
# pytest-qt>=4.2.0

# Platform-specific dependencies
# Windows: No additional requirements
# macOS: No additional requirements (uses built-in tools)
# Linux: No additional requirements

# Note: The following dependencies are NOT required for the current implementation
# as the application uses standard library modules for HTTP and file operations:
# - httpx
# - requests  
# - rapidfuzz
#
# If future versions add these dependencies, add them here:
# httpx>=0.24.0
# requests>=2.28.0
# rapidfuzz>=2.13.0
